chore: move sort_chips and get_rec_raps
