+++ recursion/tests/cached_trace.rs
@@ -1,3 +1,4 @@
+use common::VerificationParams;
 use itertools::Itertools;
 use p3_air::{Air, BaseAir};
 use p3_baby_bear::BabyBear;
@@ -13,7 +14,7 @@ use afs_stark_backend::{prover::trace::TraceCommitmentBuilder, verifier::Verific
 use afs_test_utils::config::baby_bear_poseidon2::default_engine;
 use afs_test_utils::{engine::StarkEngine, utils::generate_random_matrix};
 
-use afs_recursion::common::{self, VerificationParams};
+mod common;
 
 /// Inner value is width of y-submatrix
 pub struct SumAir(pub usize);
