+++ recursion/tests/common.rs
@@ -1,19 +1,14 @@
-use afs_compiler::ir::Config;
 use afs_stark_backend::keygen::types::MultiStarkVerifyingKey;
-use itertools::{izip, Itertools};
 use p3_baby_bear::BabyBear;
-use p3_field::PrimeField32;
 use p3_matrix::dense::RowMajorMatrix;
 use p3_matrix::Matrix;
 use p3_uni_stark::StarkGenericConfig;
 use p3_util::log2_strict_usize;
-use stark_vm::vm::ExecutionSegment;
-use std::cmp::Reverse;
 
-use crate::hints::{Hintable, InnerVal};
-use crate::stark::{DynRapForRecursion, VerifierProgram};
-use crate::types::{new_from_multi_vk, InnerConfig, VerifierInput};
 use afs_compiler::util::execute_program;
+use afs_recursion::hints::{Hintable, InnerVal};
+use afs_recursion::stark::{sort_chips, DynRapForRecursion, VerifierProgram};
+use afs_recursion::types::{new_from_multi_vk, InnerConfig, VerifierInput};
 use afs_stark_backend::prover::trace::TraceCommitmentBuilder;
 use afs_stark_backend::prover::types::Proof;
 use afs_stark_backend::rap::AnyRap;
@@ -125,51 +120,4 @@ pub fn run_recursive_test(
 
     let (program, witness_stream) = build_verification_program(rec_raps, pvs, vparams);
     execute_program::<1>(program, witness_stream);
-}
-
-#[allow(clippy::type_complexity)]
-pub fn sort_chips<'a>(
-    chips: Vec<&'a dyn AnyRap<BabyBearPoseidon2Config>>,
-    rec_raps: Vec<&'a dyn DynRapForRecursion<InnerConfig>>,
-    traces: Vec<RowMajorMatrix<BabyBear>>,
-    pvs: Vec<Vec<BabyBear>>,
-) -> (
-    Vec<&'a dyn AnyRap<BabyBearPoseidon2Config>>,
-    Vec<&'a dyn DynRapForRecursion<InnerConfig>>,
-    Vec<RowMajorMatrix<BabyBear>>,
-    Vec<Vec<BabyBear>>,
-) {
-    let mut groups = izip!(chips, rec_raps, traces, pvs).collect_vec();
-    groups.sort_by_key(|(_, _, trace, _)| Reverse(trace.height()));
-
-    let chips = groups.iter().map(|(x, _, _, _)| *x).collect_vec();
-    let rec_raps = groups.iter().map(|(_, x, _, _)| *x).collect_vec();
-    let pvs = groups.iter().map(|(_, _, _, x)| x.clone()).collect_vec();
-    let traces = groups.into_iter().map(|(_, _, x, _)| x).collect_vec();
-
-    (chips, rec_raps, traces, pvs)
-}
-
-pub fn get_rec_raps<const WORD_SIZE: usize, C: Config>(
-    vm: &ExecutionSegment<WORD_SIZE, C::F>,
-) -> Vec<&dyn DynRapForRecursion<C>>
-where
-    C::F: PrimeField32,
-{
-    let mut result: Vec<&dyn DynRapForRecursion<C>> = vec![
-        &vm.cpu_chip.air,
-        &vm.program_chip.air,
-        &vm.memory_chip.air,
-        &vm.range_checker.air,
-    ];
-    if vm.options().field_arithmetic_enabled {
-        result.push(&vm.field_arithmetic_chip.air);
-    }
-    if vm.options().field_extension_enabled {
-        result.push(&vm.field_extension_chip.air);
-    }
-    if vm.options().poseidon2_enabled() {
-        result.push(&vm.poseidon2_chip.air);
-    }
-    result
 }
