+++ recursion/tests/interactions.rs
@@ -1,6 +1,5 @@
 use std::sync::Arc;
 
-use afs_recursion::common;
 use p3_matrix::dense::RowMajorMatrix;
 
 use afs_primitives::range_gate::RangeCheckerGateChip;
@@ -12,6 +11,8 @@ use afs_test_utils::config::baby_bear_poseidon2::BabyBearPoseidon2Config;
 use afs_test_utils::config::setup_tracing;
 use afs_test_utils::interaction::dummy_interaction_air::DummyInteractionAir;
 use afs_test_utils::utils::to_field_vec;
+
+mod common;
 
 #[test]
 fn test_interactions() {
